requirements.txt
uvicorn
pyodbc
fastapi
sqlalchemy
jinja2
python-multipart
python-jose[cryptography]
passlib[bcrypt]